minimal example illustrating sphere overlap issue
Draws a box and a subset of a sphere. The sphere subset is delimited by
theta and phi limits, and its placement is rotated around the y axis by
yrot. Depending on the settings of thetamax and yrot, false overlaps
may appear. This commit has settings such that there are false overlaps.